+++ packages/bluesky/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Note
 
+## v0.4.2
+
+- Added `includeNsfw` flag to `findPopularFeed`. ([#329](https://github.com/myConsciousness/atproto.dart/issues/329))
+
 ## v0.4.1
 
 - Supported `findBlocks`. ([#320](https://github.com/myConsciousness/atproto.dart/issues/320))